Update rundeck node filter
We need to use the wpd_container_uri post meta and not the dollie()->get_container_url() function in most cases. This is because the get_container_url changes once a domain is added which results in the Rundeck node filter not working properly because that always uses the initial url as the unique identifier.

Signed-off-by: Bowe Frankema <bowe@cfcommunity.net>